collector/requirements.txt
psycopg2-binary==2.9.9
elasticsearch==8.13.0
